implement admin get total products sold endpoint

+++ src/Ecommerce.Core/Abstractions/Analytics/Services/IAnalyticsProductsService.cs
@@ -20,5 +20,5 @@ public interface IAnalyticsProductsService
 
     public Task<int> GetProductTotalViewsAsync(Guid productId, CancellationToken cancellationToken = default);
 
-    public Task<int> GetTotalProductsSoldAsync(CancellationToken cancellationToken = default);
+    public Task<long> GetTotalProductsSoldAsync(CancellationToken cancellationToken = default);
 }
